+++ TwitchLib.Api.Core/ApiBase.cs
@@ -35,45 +35,6 @@ namespace TwitchLib.Api.Core
             _jsonSerializer = new TwitchLibJsonSerializer();
         }
 
-        /// <summary>
-        /// Checks the ClientId and AccessToken against the Twitch Api Endpoints
-        /// </summary>
-        /// <returns>CredentialCheckResponseModel with a success boolean and message</returns>
-        public Task<CredentialCheckResponseModel> CheckCredentialsAsync()
-        {
-            var message = "Check successful";
-            var failMessage = "";
-            var result = true;
-            if (!string.IsNullOrWhiteSpace(Settings.ClientId) && !ValidClientId(Settings.ClientId))
-            {
-                result = false;
-                failMessage = "The passed Client Id was not valid. To get a valid Client Id, register an application here: https://www.twitch.tv/kraken/oauth2/clients/new";
-            }
-
-            if (!string.IsNullOrWhiteSpace(Settings.AccessToken) && ValidAccessToken(Settings.AccessToken) == null)
-            {
-                result = false;
-                failMessage += "The passed Access Token was not valid. To get an access token, go here:  https://twitchtokengenerator.com/";
-            }
-
-            return Task.FromResult(new CredentialCheckResponseModel { Result = result, ResultMessage = result ? message : failMessage });
-        }
-
-        public void DynamicScopeValidation(AuthScopes requiredScope, string accessToken = null)
-        {
-            //Skip validation if skip is set or access token is null
-            if (Settings.SkipDynamicScopeValidation || string.IsNullOrWhiteSpace(accessToken)) return;
-
-            //set the scopes based on access token
-            Settings.Scopes = ValidAccessToken(accessToken);
-            //skip if no scopes
-            if (Settings.Scopes == null)
-                throw new InvalidCredentialException($"The current access token does not support this call. Missing required scope: {requiredScope.ToString().ToLower()}. You can skip this check by using: IApiSettings.SkipDynamicScopeValidation = true . You can also generate a new token with this scope here: https://twitchtokengenerator.com");
-
-            if ((!Settings.Scopes.Contains(requiredScope) && requiredScope != AuthScopes.Any) || (requiredScope == AuthScopes.Any && Settings.Scopes.Any(x => x == AuthScopes.None)))
-                throw new InvalidCredentialException($"The current access token ({String.Join(",", Settings.Scopes)}) does not support this call. Missing required scope: {requiredScope.ToString().ToLower()}. You can skip this check by using: IApiSettings.SkipDynamicScopeValidation = true . You can also generate a new token with this scope here: https://twitchtokengenerator.com");
-        }
-
         internal virtual Task<Models.Root.Root> GetRootAsync(string authToken = null, string clientId = null)
         {
             return TwitchGetGenericAsync<Models.Root.Root>("", ApiVersion.V5, accessToken: authToken, clientId: clientId);
@@ -344,192 +305,6 @@ namespace TwitchLib.Api.Core
                     return propertyName.ToLower();
                 }
             }
-        }
-
-        private bool ValidClientId(string clientId)
-        {
-            try
-            {
-                var result = GetRootAsync(null, clientId).GetAwaiter().GetResult();
-                return result.Token != null;
-            }
-            catch (BadRequestException)
-            {
-                return false;
-            }
-        }
-
-        private List<AuthScopes> ValidAccessToken(string accessToken)
-        {
-            try
-            {
-                var resp = GetRootAsync(accessToken).GetAwaiter().GetResult();
-                if (resp.Token == null) return null;
-
-                return BuildScopesList(resp.Token);
-
-
-            }
-            catch
-            {
-                return null;
-            }
-        }
-
-        private static List<AuthScopes> BuildScopesList(Models.Root.RootToken token)
-        {
-            var scopes = new List<AuthScopes>();
-            foreach (var scope in token.Auth.Scopes)
-            {
-                switch (scope)
-                {
-                    case "channel_check_subscription":
-                        scopes.Add(AuthScopes.Channel_Check_Subscription);
-                        break;
-                    case "channel_commercial":
-                        scopes.Add(AuthScopes.Channel_Commercial);
-                        break;
-                    case "channel_editor":
-                        scopes.Add(AuthScopes.Channel_Editor);
-                        break;
-                    case "channel_feed_edit":
-                        scopes.Add(AuthScopes.Channel_Feed_Edit);
-                        break;
-                    case "channel_feed_read":
-                        scopes.Add(AuthScopes.Channel_Feed_Read);
-                        break;
-                    case "channel_read":
-                        scopes.Add(AuthScopes.Channel_Read);
-                        break;
-                    case "channel_stream":
-                        scopes.Add(AuthScopes.Channel_Stream);
-                        break;
-                    case "channel_subscriptions":
-                        scopes.Add(AuthScopes.Channel_Subscriptions);
-                        break;
-                    case "chat_login":
-                        scopes.Add(AuthScopes.Chat_Login);
-                        break;
-                    case "collections_edit":
-                        scopes.Add(AuthScopes.Collections_Edit);
-                        break;
-                    case "communities_edit":
-                        scopes.Add(AuthScopes.Communities_Edit);
-                        break;
-                    case "communities_moderate":
-                        scopes.Add(AuthScopes.Communities_Moderate);
-                        break;
-                    case "user_blocks_edit":
-                        scopes.Add(AuthScopes.User_Blocks_Edit);
-                        break;
-                    case "user_blocks_read":
-                        scopes.Add(AuthScopes.User_Blocks_Read);
-                        break;
-                    case "user_follows_edit":
-                        scopes.Add(AuthScopes.User_Follows_Edit);
-                        break;
-                    case "user_read":
-                        scopes.Add(AuthScopes.User_Read);
-                        break;
-                    case "user_subscriptions":
-                        scopes.Add(AuthScopes.User_Subscriptions);
-                        break;
-                    case "openid":
-                        scopes.Add(AuthScopes.OpenId);
-                        break;
-                    case "viewing_activity_read":
-                        scopes.Add(AuthScopes.Viewing_Activity_Read);
-                        break;
-                    case "user:edit:broadcast":
-                        scopes.Add(AuthScopes.Helix_User_Edit_Broadcast);
-                        break;
-                    case "analytics:read:extensions":
-                        scopes.Add(AuthScopes.Helix_Analytics_Read_Extensions);
-                        break;
-                    case "analytics:read:games":
-                        scopes.Add(AuthScopes.Helix_Analytics_Read_Games);
-                        break;
-                    case "bits:read":
-                        scopes.Add(AuthScopes.Helix_Bits_Read);
-                        break;
-                    case "channel:edit:commercial":
-                        scopes.Add(AuthScopes.Helix_Channel_Edit_Commercial);
-                        break;
-                    case "channel:manage:broadcast":
-                        scopes.Add(AuthScopes.Helix_Channel_Manage_Broadcast);
-                        break;
-                    case "channel:manage:extensions":
-                        scopes.Add(AuthScopes.Helix_Channel_Manage_Extensions);
-                        break;
-                    case "channel:manage:redemptions":
-                        scopes.Add(AuthScopes.Helix_Channel_Manage_Redemptions);
-                        break;
-                    case "channel:read:hype_train":
-                        scopes.Add(AuthScopes.Helix_Channel_Read_Hype_Train);
-                        break;
-                    case "channel:read:redemptions":
-                        scopes.Add(AuthScopes.Helix_Channel_Read_Redemptions);
-                        break;
-                    case "channel:read:stream_key":
-                        scopes.Add(AuthScopes.Helix_Channel_Read_Stream_Key);
-                        break;
-                    case "channel:read:subscriptions":
-                        scopes.Add(AuthScopes.Helix_Channel_Read_Subscriptions);
-                        break;
-                    case "clips:edit":
-                        scopes.Add(AuthScopes.Helix_Clips_Edit);
-                        break;
-                    case "moderation:read":
-                        scopes.Add(AuthScopes.Helix_Moderation_Read);
-                        break;
-                    case "user:edit":
-                        scopes.Add(AuthScopes.Helix_User_Edit);
-                        break;
-                    case "user:edit:follows":
-                        scopes.Add(AuthScopes.Helix_User_Edit_Follows);
-                        break;
-                    case "user:read:broadcast":
-                        scopes.Add(AuthScopes.Helix_User_Read_Broadcast);
-                        break;
-                    case "user:read:email":
-                        scopes.Add(AuthScopes.Helix_User_Read_Email);
-                        break;
-                    case "channel:read:editors":
-                        scopes.Add(AuthScopes.Helix_Channel_Read_Editors);
-                        break;
-                    case "channel:manage:videos":
-                        scopes.Add(AuthScopes.Helix_Channel_Manage_Videos);
-                        break;
-                    case "user:read:blocked_users":
-                        scopes.Add(AuthScopes.Helix_User_Read_BlockedUsers);
-                        break;
-                    case "user:manage:blocked_users":
-                        scopes.Add(AuthScopes.Helix_User_Manage_BlockedUsers);
-                        break;
-                    case "user:read:subscriptions":
-                        scopes.Add(AuthScopes.Helix_User_Read_Subscriptions);
-                        break;
-                    case "channel:manage:polls":
-                        scopes.Add(AuthScopes.Helix_Channel_Manage_Polls);
-                        break;
-                    case "channel:manage:predictions":
-                        scopes.Add(AuthScopes.Helix_Channel_Manage_Predictions);
-                        break;
-                    case "channel:read:polls":
-                        scopes.Add(AuthScopes.Helix_Channel_Read_Polls);
-                        break;
-                    case "channel:read:predictions":
-                        scopes.Add(AuthScopes.Helix_Channel_Read_Predictions);
-                        break;
-                    case "moderator:manage:automod":
-                        scopes.Add(AuthScopes.Helix_Channel_Moderator_Manage_Automod);
-                        break;
-                }
-            }
-
-            if (scopes.Count == 0)
-                scopes.Add(AuthScopes.None);
-            return scopes;
         }
 
         private string ConstructResourceUrl(string resource = null, List<KeyValuePair<string, string>> getParams = null, ApiVersion api = ApiVersion.V5, string overrideUrl = null)
