+++ TwitchLib.Api.V5/Clips.cs
@@ -67,7 +67,6 @@ namespace TwitchLib.Api.V5
 
         public Task<FollowClipsResponse> GetFollowedClipsAsync(long limit = 10, string cursor = null, bool trending = false, string authToken = null)
         {
-            DynamicScopeValidation(AuthScopes.User_Read, authToken);
             var getParams = new List<KeyValuePair<string, string>>
                 {
                     new KeyValuePair<string, string>("limit", limit.ToString())
