+++ TwitchLib.Api.V5/Collections.cs
@@ -66,7 +66,6 @@ namespace TwitchLib.Api.V5
 
         public Task<CollectionMetadata> CreateCollectionAsync(string channelId, string collectionTitle, string authToken = null)
         {
-            DynamicScopeValidation(AuthScopes.Collections_Edit, authToken);
             if (string.IsNullOrWhiteSpace(channelId))
                 throw new BadParameterException("The channel id is not valid for a collection creation. It is not allowed to be null, empty or filled with whitespaces.");
 
@@ -83,7 +82,6 @@ namespace TwitchLib.Api.V5
 
         public Task UpdateCollectionAsync(string collectionId, string newCollectionTitle, string authToken = null)
         {
-            DynamicScopeValidation(AuthScopes.Collections_Edit, authToken);
             if (string.IsNullOrWhiteSpace(collectionId))
                 throw new BadParameterException("The collection id is not valid for a collection. It is not allowed to be null, empty or filled with whitespaces.");
 
@@ -100,7 +98,6 @@ namespace TwitchLib.Api.V5
 
         public Task CreateCollectionThumbnailAsync(string collectionId, string itemId, string authToken = null)
         {
-            DynamicScopeValidation(AuthScopes.Collections_Edit, authToken);
             if (string.IsNullOrWhiteSpace(collectionId))
                 throw new BadParameterException("The collection id is not valid for a collection. It is not allowed to be null, empty or filled with whitespaces.");
 
@@ -117,7 +114,6 @@ namespace TwitchLib.Api.V5
 
         public Task DeleteCollectionAsync(string collectionId, string authToken = null)
         {
-            DynamicScopeValidation(AuthScopes.Collections_Edit, authToken);
             if (string.IsNullOrWhiteSpace(collectionId))
                 throw new BadParameterException("The collection id is not valid for a collection. It is not allowed to be null, empty or filled with whitespaces.");
 
@@ -130,7 +126,6 @@ namespace TwitchLib.Api.V5
 
         public Task<CollectionItem> AddItemToCollectionAsync(string collectionId, string itemId, string itemType, string authToken = null)
         {
-            DynamicScopeValidation(AuthScopes.Collections_Edit, authToken);
             if (string.IsNullOrWhiteSpace(collectionId))
                 throw new BadParameterException("The collection id is not valid for a collection. It is not allowed to be null, empty or filled with whitespaces.");
 
@@ -150,7 +145,6 @@ namespace TwitchLib.Api.V5
 
         public Task DeleteItemFromCollectionAsync(string collectionId, string itemId, string authToken = null)
         {
-            DynamicScopeValidation(AuthScopes.Collections_Edit, authToken);
             if (string.IsNullOrWhiteSpace(collectionId))
                 throw new BadParameterException("The collection id is not valid for a collection. It is not allowed to be null, empty or filled with whitespaces.");
 
@@ -166,7 +160,6 @@ namespace TwitchLib.Api.V5
 
         public Task MoveItemWithinCollectionAsync(string collectionId, string itemId, int position, string authToken = null)
         {
-            DynamicScopeValidation(AuthScopes.Collections_Edit, authToken);
             if (string.IsNullOrWhiteSpace(collectionId))
                 throw new BadParameterException("The collection id is not valid for a collection. It is not allowed to be null, empty or filled with whitespaces.");
